YT-29420: Faster SetLogicalType commit_hash:399da7a93b6f26bfacd27fcc7dfedcc748252226
