+++ yt/yt/client/table_client/private.h
@@ -12,4 +12,16 @@ YT_DEFINE_LEAKY_GLOBAL(const NLogging::TLogger, TableClientLogger, "TableClient"
 
 ////////////////////////////////////////////////////////////////////////////////
 
+struct TTypeV3Info
+{
+    ESimpleLogicalValueType V1Type;
+    EValueType WireType;
+    bool Required;
+    bool IsPureV1Type;
+};
+
+TTypeV3Info GetTypeV3Info(const TLogicalTypePtr& logicalType);
+
+////////////////////////////////////////////////////////////////////////////////
+
 } // namespace NYT::NTableClient
